+++ app/routes/contacts.$contactId.destroy.tsx
@@ -7,5 +7,5 @@ import { deleteContact } from '../data'
 export const action = async ({ params }: ActionFunctionArgs) => {
   invariant(params.contactId, 'Missing contactId param')
   await deleteContact(params.contactId)
-  return redirect('/')
+  return redirect('/contacts')
 }
